fix(hle): support current dev firmware interfaces
Migrate the emulator to categorized 16-bit SPI commands and the five-byte
frame header. Exclude hardware-only companion transports while retaining
explicit HLE unsupported adapters, and enable the QR widget required by the
current UI.

+++ tools/hle/src/hal/hle_kernel.cpp
@@ -127,7 +127,7 @@ static std::thread s_bridge_worker;
 static void c5_bridge_worker(hle::SPIBridgeChannel *channel) {
     ESP_LOGI(TAG, "C5 bridge worker started");
     while (!channel->is_closed()) {
-        uint8_t cmd_id;
+        uint16_t cmd_id;
         uint8_t payload[256];
         uint8_t payload_len;
         if (!channel->slave_wait_command(cmd_id, payload, payload_len)) {
